+++ src/features/article-analyze/api/getAnalysisById.ts
@@ -1,0 +1,13 @@
+import { createClient } from '@/shared/lib/supabase/client';
+
+export async function getAnalysisById(id: string) {
+  const supabase = await createClient();
+
+  const { data, error } = await supabase
+    .from('analyses')
+    .select('*')
+    .eq('id', id)
+    .single();
+
+  return { data, error };
+}
