+++ src/features/article-analyze/index.ts
@@ -6,11 +6,15 @@ export { ArticleCard } from './ui/ArticleCard';
 export { CopyButton } from './ui/CopyButton';
 export { RecentAnalysesList } from './ui/RecentAnalysesList';
 export { ToggleButton } from './ui/ToggleButton';
-export {OriginalLinkButton} from './ui/OriginalLinkButton';
+export { OriginalLinkButton } from './ui/OriginalLinkButton';
+export { ArticleAnalyzeSection } from './ui/ArticleAnalyzeSection';
+export { ArticleAnalyzeHeader } from './ui/ArticleAnalyzeHeader';
+export { AnalysisDescription } from './ui/AnalysisDescription';
+export { SingleAnalysisSection } from './ui/SingleAnalysisSection';
+export { PairAnalysisSection } from './ui/PairAnalysisSection';
 
 //api
-export { useNewsSearch } from './api/useNewsSearch';
-export { useAnalyze } from './api/useAnalyze';
+export { getAnalysisById } from './api/getAnalysisById';
 
 // model
 export type { NewsResultProps, NewsItem } from './model/type';
@@ -20,7 +24,9 @@ export { useAnalyzeModel } from './model/useAnalyzeModel';
 export { useSelectedNewsStore } from './model/useSelectedNewsStore';
 export type { Article, ResultPageProps } from './model/type';
 export { useToastMessageStore } from './model/useToastMessageStore';
-export { useExperimentLog } from './api/useExperimentLog';
+export { logExperiment } from './api/logExperiment';
+export { useAnalyze } from './model/useAnalyze';
+export { useNewsSearch } from './model/useNewsSearch';
 
 //lib
 export { getSourceName } from './lib/newspaperFormat';
@@ -28,3 +34,6 @@ export { stripHtml } from './lib/striphtml';
 export { dateFormat } from './lib/dateFormat';
 export { TONE_STYLE } from '../article-analyze/lib/constants';
 export { ANALYSIS_ROW_KEYS } from '../article-analyze/lib/constants';
+export { buildRows } from './lib/buildRows';
+export { buildCommonKeywords } from './lib/buildRows';
+export { buildAnalysisMetadata } from './lib/buildMetadata';
