+++ src/features/article-analyze/model/useAnalyze.ts
@@ -2,8 +2,8 @@
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import axios from 'axios';
-import { AnalyzeParams } from '../model/type';
-import { useToastMessageStore } from '../model/useToastMessageStore';
+import { AnalyzeParams } from './type';
+import { useToastMessageStore } from './useToastMessageStore';
 
 export function useAnalyze() {
   const router = useRouter();
@@ -19,7 +19,6 @@ export function useAnalyze() {
     },
     onError: (error) => {
       if (axios.isAxiosError(error)) {
-        console.log('Axios error:', error.response?.data || error.message);
         setToastMessage('분석 중 오류가 발생했습니다. 다시 시도해주세요.');
       }
     },
