+++ src/features/article-analyze/model/useAnalyzeModel.ts
@@ -1,13 +1,13 @@
 'use client';
-import { useAnalyze } from '../api/useAnalyze';
-import { useExperimentLog } from '../api/useExperimentLog';
+import { useAnalyze } from './useAnalyze';
+import { logExperiment } from '../api/logExperiment';
 import { getSourceName } from '../lib/newspaperFormat';
 import { stripHtml } from '../lib/striphtml';
 import type { NewsItem, SearchMode } from './type';
 
 export function useAnalyzeModel(mode: SearchMode, query: string) {
   const { mutate: analyze, isPending, isError: isAnalyzeError } = useAnalyze();
-  const { log } = useExperimentLog();
+  const { log } = logExperiment();
 
   const handleCompare = (selected: NewsItem[], keyword: string) => {
     log({ mode, query, eventType: 'compare_start' }); // 비교 시작 로그
