+++ src/features/article-analyze/model/useCluster.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import axios from 'axios';
-import { NaverArticle, ClusterResult } from '../model/type';
-import { useToastMessageStore } from '../model/useToastMessageStore';
+import { NaverArticle, ClusterResult } from './type';
+import { useToastMessageStore } from './useToastMessageStore';
 
 export function useCluster() {
   const { setToastMessage } = useToastMessageStore();
@@ -9,11 +9,12 @@ export function useCluster() {
   return useMutation<ClusterResult, Error, NaverArticle[]>({
     mutationFn: async (items: NaverArticle[]) => {
       const { data } = await axios.post('/api/cluster', { items });
+
       return data;
     },
     onError: (error) => {
       console.error('클러스터링 중 오류 발생:', error);
-      if(axios.isAxiosError(error)) {
+      if (axios.isAxiosError(error)) {
         setToastMessage('클러스터링 중 오류가 발생했습니다.');
       }
     },
