--- src/widgets/analyze-entry/index.ts
@@ -1,1 +1,0 @@
-export {ArticleAnalyzeSection} from "./ui/ArticleAnalyzeSection"
